Add files via upload
For context, the file retrieved and uploaded was a backup from February 21, 2022.
This one is also from February of 2022, but it was the 20th. We think it got corrupted from the service immediately closing.
The website stayed up till the 22nd and then closed but it allowed for downloads of animations.